refactor: switch to relative paths and clean up folder structure
