+++ model/Custom_LipForensics/LipForensics/inference.py
@@ -13,19 +13,15 @@ from torchvision.transforms import Compose, CenterCrop, Resize, Lambda
 from tqdm import tqdm
 
 import sys
-sys.path.insert(0, "./data")
+sys.path.insert(0, "./LipForensics/data")
 from transforms import NormalizeVideo, ToTensorVideo
 from dataset_clips import SingleVideoClips
-from samplers import ConsecutiveClipSampler
-sys.path.insert(0, "./models")
+sys.path.insert(0, "./LipForensics/models")
 from spatiotemporal_net import get_model
-#sys.path.insert(0, '/root/25th-conference-fakebusters/model/Custom_LipForensics/LipForensics')
-from lipforensics_utils import get_files_from_split
-
 
 def evaluate_lipforensics(
     cropped_mouths_array,
-    weights_forgery_path="./models/weights/lipforensics_ff.pth",
+    weights_forgery_path="./LipForensics/models/weights/lipforensics_ff.pth",
     frames_per_clip=25,
     batch_size=8,
     device="cuda:0",
