+++ model/Custom_LipForensics/README.md
@@ -1,2 +1,4 @@
 ## How to run
-`python final_inference_roi.py --video_path /root/chimchak.mp4 --config_path /root/25th-conference-fakebusters/Custom_LipForensics/roi_extractor/config.yaml`
+`python run_dlib.py --video_path ./video.mp4`
+
+The result indicates the probability that the input video is fake.
