+++ model/Custom_LipForensics/roi_extractor/dlib/requirements.txt
@@ -1,0 +1,6 @@
+tqdm==4.67.1
+pyyaml==6.0.2
+dlib==19.24.6
+scikit-image==0.24.0
+scikit-video==1.1.11
+opencv-python==4.10.0.84
